# perf(cli): dynamically import `resolveConfig`