--- a/packages/vite/src/node/cli.ts
+++ b/packages/vite/src/node/cli.ts
@@ -9,7 +9,6 @@ import type { ServerOptions } from './server'
 import type { CLIShortcut } from './shortcuts'
 import type { LogLevel } from './logger'
 import { createLogger } from './logger'
-import { resolveConfig } from './config'
 import type { InlineConfig } from './config'
 
 const cli = cac('vite')
@@ -352,6 +351,7 @@ cli
   .action(
     async (root: string, options: { force?: boolean } & GlobalCLIOptions) => {
       filterDuplicateOptions(options)
+      const { resolveConfig } = await import('./config')
       const { optimizeDeps } = await import('./optimizer')
       try {
         const config = await resolveConfig(