### Description

I noticed that `resolveConfig` was top-level imported but only used in `vite optimize` in the CLI entrypoint. Making it a dynamic import doesn't change much for normal `vite dev` or `vite build` calls, but should help with `vite --version` or `vite --help`.

With `time vite --help`, before:

```
0.12s user 0.02s system 133% cpu 0.108 total
```

after:

```
0.04s user 0.01s system 91% cpu 0.049 total
```

<!-- What is this PR solving? Write a clear description or reference the issues it solves (e.g. `fixes #123`). What other alternatives have you explored? Are there any parts you think require more attention from reviewers? -->

<!----------------------------------------------------------------------
Before creating the pull request, please make sure you do the following:

- Read the Contributing Guidelines at https://github.com/vitejs/vite/blob/main/CONTRIBUTING.md.
- Check that there isn't already a PR that solves the problem the same way. If you find a duplicate, please help us reviewing it.
- Update the corresponding documentation if needed.
- Include relevant tests that fail without this PR but pass with it.

Thank you for contributing to Vite!
----------------------------------------------------------------------->